+++ package.js
@@ -1,6 +1,6 @@
 Package.describe({
     name: 'pwix:roles',
-    version: '1.6.3-rc',
+    version: '1.7.0-rc',
     summary: 'An encapsulation of alanning:roles',
     git: 'https://github.com/trychlos/pwix-roles',
     documentation: 'README.md'
@@ -29,12 +29,14 @@ function configure( api ){
     api.use( 'ecmascript' );
     api.use( 'less@4.0.0', 'client' );
     api.use( 'mongo', 'server' );
+    api.use( 'pwix:accounts-hub@1.0.0' );
     api.use( 'pwix:bootbox@1.5.0' );
     api.use( 'pwix:i18n@1.5.7' );
     api.use( 'pwix:jstree@1.0.6' );
     api.use( 'pwix:modal@1.10.0 || 2.0.0' );
     api.use( 'pwix:plus-button@1.0.0' );
     api.use( 'pwix:tabbed@1.0.0-rc' );
+    api.use( 'pwix:tolert@1.5.0' );
     api.use( 'pwix:ui-bootstrap5@2.0.0' );
     api.use( 'pwix:ui-fontawesome6@1.0.0' );
     api.use( 'pwix:ui-utils@1.0.0' );
@@ -42,6 +44,7 @@ function configure( api ){
     api.use( 'reactive-var' );
     api.use( 'tmeasday:check-npm-versions@1.0.2 || 2.0.0-beta.0', 'server' );
     api.use( 'tracker' );
+    api.addFiles( 'src/client/components/prScopedAccountsPanel/prScopedAccountsPanel.js', 'client' );
     api.addFiles( 'src/client/components/prEdit/prEdit.js', 'client' );
     api.addFiles( 'src/client/components/prEditPanel/prEditPanel.js', 'client' );
     api.addFiles( 'src/client/components/prView/prView.js', 'client' );