+++ src/client/components/pr_edit_scoped_plus/pr_edit_scoped_plus.js
@@ -1,0 +1,10 @@
+/*
+ * pwix:roles/src/client/components/pr_edit_scoped_plus/pr_edit_scoped_plus.js
+ *
+ *  Have a 'New scope' button
+ * 
+ *  Parms:
+ *  - the PlusButton parameters, provided by prEditPanel
+ */
+
+import './pr_edit_scoped_plus.html';